テストを書きやすくするように dll のロード処理を外に出した
