+++ src/aquestalk.rs
@@ -15,9 +15,12 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with AquesTalk-proxy.  If not, see <https://www.gnu.org/licenses/>.
 
+use std::collections::HashMap;
 use std::ffi::OsStr;
 use std::fmt;
+use std::fs;
 use std::ops::Deref;
+use std::path::Path;
 use std::slice;
 use std::sync::{Arc, Mutex};
 
@@ -136,4 +139,21 @@ impl Drop for Wav {
             dll.free_wave(self.wav);
         }
     }
+}
+
+pub fn load_libs<P>(path: &P) -> Result<HashMap<String, AquesTalk>, Box<dyn std::error::Error>>
+where
+    P: AsRef<Path>,
+{
+    let mut aqtks = HashMap::new();
+    for entry in fs::read_dir(path)? {
+        let entry = entry?;
+        if entry.file_type()?.is_dir() {
+            let voice_type = entry.file_name().into_string().unwrap();
+            let mut path = entry.path();
+            path.push("AquesTalk.dll");
+            aqtks.insert(voice_type, AquesTalk::new(path.into_os_string())?);
+        }
+    }
+    Ok(aqtks)
 }
